refactor(link): extract app picker into shared lib/app-picker
Pull the 'pick existing or create new' interactive flow out of
commands/link/index.ts into lib/app-picker.ts so other entry points can
use the same UX. No behavior change for clerk link itself; the picker
message and the auto-detect-from-keys handshake stay where they were.
